feat: CaT complete
Note that this channel reads calcium concentration but does not
contribute to it, producing a different current density, much like a
calcium dependent potassium channel.

So, we use a different ion here to distinguish it from the "standard"
calcium ion. When using this channel, we need to remember to set this
particular ion in the GHK channel density also.